Create NeighborJoining
A recursive neighbor-joining algorithm for generating phylogenetic trees from additive and non-additive distance matrices. The inputs are an integer n and a space-separated n by n distance matrix. The matrix is converted into a dictionary of dictionaries. The output is an adjacency list describing the phylogenetic tree, also in the form of a dictionary of dictionaries.